+++ lib/src/element/sub/button/radio_btn_element.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-
+import '../../../element/super/multi_child_element.dart';
 import '../../../element_params/element_child.dart';
 import '../../../element_params/sub/button/checkbox_params.dart';
 import '../../../element_params/sub/button/radio_btn_params.dart';
@@ -7,15 +7,13 @@ import '../../../element_params/super/spwml_params.dart';
 import '../../../enum/enum_spwml_params.dart';
 import '../../../enum/enum_spwml_element_type.dart';
 import '../../../style/spwml_font_style.dart';
-import '../../super/spwml_element.dart';
 
 ///
 /// Author Masahide Mori
 ///
 /// First edition creation date 2022-06-13 21:13:22
 ///
-class RadioBtnElement extends SpWMLElement {
-  final StructureElementChildren children;
+class RadioBtnElement extends MultiChildElement {
   final RadioBtnParamsWrapper elParams;
 
   ///
@@ -40,10 +38,10 @@ class RadioBtnElement extends SpWMLElement {
       int lineStart,
       int lineEnd,
       SpWMLFontStyle style,
-      this.children,
+      StructureElementChildren children,
       this.elParams)
       : super(serial, EnumSpWMLElementType.radioBtn, params, spwmlParams,
-            parentSerial, lineStart, lineEnd, style);
+            parentSerial, lineStart, lineEnd, style, children);
 
   @override
   RadioBtnElement initParams() {