+++ lib/src/element/sub/structure/exp_tile_element.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import '../../super/multi_child_text_element.dart';
 import '../../../element_params/element_child.dart';
 import '../../../element_params/sub/structure/exp_tile_params.dart';
 import '../../../element_params/sub/text/text_params.dart';
@@ -6,15 +7,13 @@ import '../../../element_params/super/spwml_params.dart';
 import '../../../enum/enum_spwml_params.dart';
 import '../../../enum/enum_spwml_element_type.dart';
 import '../../../style/spwml_font_style.dart';
-import '../text/text_element.dart';
 
 ///
 /// Author Masahide Mori
 ///
 /// First edition creation date 2021-12-31 17:58:41
 ///
-class ExpTileElement extends TextElement {
-  final StructureElementChildren children;
+class ExpTileElement extends MultiChildTextElement {
   final ExpTileParamsWrapper elParams;
 
   ///
@@ -39,11 +38,11 @@ class ExpTileElement extends TextElement {
       int lineStart,
       int lineEnd,
       SpWMLFontStyle style,
-      this.children,
+      StructureElementChildren children,
       TextParamsWrapper textParams,
       this.elParams)
       : super(serial, EnumSpWMLElementType.expTile, params, spwmlParams,
-            parentSerial, lineStart, lineEnd, style, textParams);
+            parentSerial, lineStart, lineEnd, style, children, textParams);
 
   @override
   ExpTileElement initParams() {