+++ lib/src/element/sub/structure/stack_element.dart
@@ -1,18 +1,17 @@
 import 'package:flutter/material.dart';
+import '../../../element/spwml_element_part.dart';
 import '../../../element_params/element_child.dart';
 import '../../../element_params/sub/structure/stack_params.dart';
 import '../../../element_params/super/spwml_params.dart';
 import '../../../enum/enum_spwml_element_type.dart';
 import '../../../style/spwml_font_style.dart';
-import '../../super/spwml_element.dart';
 
 ///
 /// Author Masahide Mori
 ///
 /// First edition creation date 2022-01-06 14:53:11
 ///
-class StackElement extends SpWMLElement {
-  final StructureElementChildren children;
+class StackElement extends MultiChildElement {
   final StackParamsWrapper elParams;
 
   ///
@@ -37,10 +36,10 @@ class StackElement extends SpWMLElement {
       int lineStart,
       int lineEnd,
       SpWMLFontStyle style,
-      this.children,
+      StructureElementChildren children,
       this.elParams)
       : super(serial, EnumSpWMLElementType.stack, params, spwmlParams,
-            parentSerial, lineStart, lineEnd, style);
+            parentSerial, lineStart, lineEnd, style, children);
 
   @override
   StackElement initParams() {