+++ lib/src/element/sub/structure/wrap_element.dart
@@ -1,19 +1,18 @@
 import 'package:flutter/material.dart';
+import '../../../element/super/multi_child_element.dart';
 import '../../../element_params/element_child.dart';
 import '../../../element_params/sub/structure/wrap_params.dart';
 import '../../../element_params/super/spwml_params.dart';
 import '../../../enum/enum_spwml_params.dart';
 import '../../../enum/enum_spwml_element_type.dart';
 import '../../../style/spwml_font_style.dart';
-import '../../super/spwml_element.dart';
 
 ///
 /// Author Masahide Mori
 ///
 /// First edition creation date 2022-04-10 16:18:52
 ///
-class WrapElement extends SpWMLElement {
-  final StructureElementChildren children;
+class WrapElement extends MultiChildElement {
   final WrapParamsWrapper elParams;
 
   ///
@@ -38,10 +37,10 @@ class WrapElement extends SpWMLElement {
       int lineStart,
       int lineEnd,
       SpWMLFontStyle style,
-      this.children,
+      StructureElementChildren children,
       this.elParams)
       : super(serial, EnumSpWMLElementType.wrap, params, spwmlParams,
-            parentSerial, lineStart, lineEnd, style);
+            parentSerial, lineStart, lineEnd, style, children);
 
   @override
   WrapElement initParams() {